Extract shared utilities from data_source_query
